DAOS-18231 test: Move tests dfuse/daos_build_vm.py to daos_build.py
Move the tests from VM to HW Medium based on the
triage meeting. Remove daos_build_vm.py and yaml.

Skip-unit-tests: true
Skip-fault-injection-test: true
Skip-func-hw-test-medium: false
Test-tag: test_dfuse_daos_build_wt_il test_dfuse_daos_build_wt_pil4dfs
