+++ src/tests/ftest/dfuse/daos_build.py
@@ -306,3 +306,35 @@ class DaosBuild(TestWithServers):
         :avocado: tags=DaosBuild,test_dfuse_daos_build_nocache
         """
         run_build_test(self, "nocache")
+
+    def test_dfuse_daos_build_wt_il(self):
+        """This test builds DAOS on a dfuse filesystem.
+
+        Use cases:
+            Create Pool
+            Create Posix container
+            Mount dfuse
+            Checkout and build DAOS sources.
+
+        :avocado: tags=all,full_regression
+        :avocado: tags=hw,medium
+        :avocado: tags=daosio,dfs,dfuse,ioil
+        :avocado: tags=DaosBuild,test_dfuse_daos_build_wt_il
+        """
+        run_build_test(self, "writethrough", il_lib='libioil.so', run_on_vms=True)
+
+    def test_dfuse_daos_build_wt_pil4dfs(self):
+        """This test builds DAOS on a dfuse filesystem.
+
+        Use cases:
+            Create Pool
+            Create Posix container
+            Mount dfuse
+            Checkout and build DAOS sources.
+
+        :avocado: tags=all,full_regression
+        :avocado: tags=hw,medium
+        :avocado: tags=daosio,dfs,dfuse,pil4dfs
+        :avocado: tags=DaosBuild,test_dfuse_daos_build_wt_pil4dfs
+        """
+        run_build_test(self, "nocache", il_lib='libpil4dfs.so', run_on_vms=True)
